Move incident-related elements into one folder.
